Move connection to sql package
